Fix all unit tests related to environment updates
