+++ src/components/ApplicationListView/__tests__/ApplicationListRow.spec.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import { useK8sWatchResource } from '@openshift/dynamic-plugin-sdk-utils';
 import { render } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { mockAppEnvWithHealthStatus } from '../../../hacbs/components/Environment/__data__/mockAppEnvWithHealthStatus';
 import { useAllEnvironments } from '../../../hooks/useAllEnvironments';
 import { useSnapshotsEnvironmentBindings } from '../../../hooks/useSnapshotsEnvironmentBindings';
 import * as dateTime from '../../../shared/components/timestamp/datetime';
 import { ApplicationKind, ComponentKind } from '../../../types';
 import { mockSnapshotsEnvironmentBindings } from '../../ApplicationDetails/__data__';
+import { mockAppEnvWithHealthStatus } from '../../Environment/__data__/mockAppEnvWithHealthStatus';
 import ApplicationListRow from '../ApplicationListRow';
 
 jest.mock('@openshift/dynamic-plugin-sdk-utils', () => ({
