+++ src/components/Environment/__data__/mockAppEnvWithHealthStatus.ts
@@ -68,6 +68,7 @@ export const mockAppEnvWithHealthStatus = [
       displayName: 'Production',
       parentEnvironment: 'staging',
       type: 'non-poc',
+      tags: ['static'],
     },
     healthStatus: 'Missing',
   },
@@ -105,6 +106,7 @@ export const mockAppEnvWithHealthStatus = [
       displayName: 'Staging',
       parentEnvironment: 'development',
       type: 'poc',
+      tags: ['static'],
     },
     healthStatus: 'Missing',
   },
@@ -122,6 +124,7 @@ export const mockAppEnvWithHealthStatus = [
       deploymentStrategy: 'Manual',
       displayName: 'SRE environment',
       type: 'poc',
+      tags: ['managed'],
     },
     healthStatus: 'Healthy',
   },
