+++ src/components/Environment/__tests__/EnvironmentCard.spec.tsx
@@ -1,32 +1,19 @@
-import * as React from 'react';
-import '@testing-library/jest-dom';
+import React from 'react';
 import { useFeatureFlag } from '@openshift/dynamic-plugin-sdk';
-import { useK8sWatchResource } from '@openshift/dynamic-plugin-sdk-utils';
-import { render, screen, configure } from '@testing-library/react';
-import { useApplications } from '../../../hooks/useApplications';
+import '@testing-library/jest-dom';
+import { getActiveWorkspace } from '@openshift/dynamic-plugin-sdk-utils';
+import { render, screen } from '@testing-library/react';
+import { EnvironmentModel } from '../../../models';
 import { EnvironmentKind } from '../../../types';
-import { mockApplication } from '../../ApplicationDetails/__data__/mock-data';
 import EnvironmentCard from '../EnvironmentCard';
-
-jest.mock('@openshift/dynamic-plugin-sdk-utils', () => ({
-  useK8sWatchResource: jest.fn(),
-  getActiveWorkspace: jest.fn(() => 'test-ws'),
-}));
-
-jest.mock('react-i18next', () => ({
-  useTranslation: jest.fn(() => ({ t: (x) => x })),
-}));
-
-jest.mock('react-router-dom', () => ({
-  Link: (props) => <a href={props.to}>{props.children}</a>,
-}));
-
-jest.mock('../../../hooks/useApplications', () => ({
-  useApplications: jest.fn(),
-}));
 
 jest.mock('@openshift/dynamic-plugin-sdk', () => ({
   useFeatureFlag: jest.fn(),
+}));
+
+jest.mock('@openshift/dynamic-plugin-sdk-utils', () => ({
+  commonFetch: jest.fn(),
+  getActiveWorkspace: jest.fn(),
 }));
 
 jest.mock('../../../utils/rbac', () => ({
@@ -36,45 +23,56 @@ jest.mock('../../../utils/rbac', () => ({
 const useFeatureFlagMock = useFeatureFlag as jest.Mock;
 useFeatureFlagMock.mockReturnValue([false, () => {}]);
 
-const environment: EnvironmentKind = {
-  kind: 'Environment',
-  apiVersion: 'appstudio.redhat.com/v1alpha1',
+const getActiveWorkspaceMock = getActiveWorkspace as jest.Mock;
+getActiveWorkspaceMock.mockReturnValue('test-ws');
+
+const testEnv: EnvironmentKind = {
+  apiVersion: `${EnvironmentModel.apiGroup}/${EnvironmentModel.apiVersion}`,
+  kind: EnvironmentModel.kind,
   metadata: {
-    name: 'prod',
-    namespace: 'abhi',
-    resourceVersion: '187593762',
-    uid: '60725777-a545-4c54-bf25-19a3f231aed1',
+    name: 'test',
   },
   spec: {
-    displayName: 'Prod',
-    clusterCredentials: {
-      apiServerURL: 'link1',
-    },
-    deploymentStrategy: 'Manual',
+    displayName: 'My Env',
     type: 'poc',
+    deploymentStrategy: 'Manual',
+    tags: ['static'],
   },
 };
-configure({ testIdAttribute: 'data-test' });
 
-const watchResourceMock = useK8sWatchResource as jest.Mock;
-
-const useApplicationsMock = useApplications as jest.Mock;
-
-describe('EnvironmentCard', () => {
-  it('should render correct CardBody sections', () => {
-    watchResourceMock.mockReturnValue([[], true]);
-    useApplicationsMock.mockReturnValue([[mockApplication], true]);
-    render(<EnvironmentCard environment={environment} />);
-    screen.getByText(environment.spec.displayName);
-    screen.getByText('Applications: 1');
+describe('', () => {
+  it('should render environment card', () => {
+    render(<EnvironmentCard environment={testEnv} />);
+    expect(screen.getByText('Static')).toBeVisible();
+    expect(screen.getByText(testEnv.spec.displayName)).toBeVisible();
+    expect(screen.queryByText(testEnv.metadata.name)).toBeNull();
+    expect(screen.getByText('Manual')).toBeVisible();
   });
 
-  it('should render correct Deployment strategy CardBody sections', () => {
-    watchResourceMock.mockReturnValue([[], true]);
-    environment.spec.deploymentStrategy = 'AppStudioAutomated';
-    render(<EnvironmentCard environment={environment} />);
-    screen.getByText(environment.spec.displayName);
-    screen.getByText('Deployment strategy:');
-    screen.getByText('Automatic');
+  it('should render environment metadata.name when displayName is absent', () => {
+    const env = {
+      ...testEnv,
+      spec: {
+        ...testEnv.spec,
+        displayName: undefined,
+      },
+    };
+    render(<EnvironmentCard environment={env} />);
+    expect(screen.getByText(testEnv.metadata.name)).toBeVisible();
+    expect(screen.queryByText(testEnv.spec.displayName)).toBeNull();
+  });
+
+  it('should render Application Healthy card in the list view', () => {
+    const env = {
+      ...testEnv,
+      spec: {
+        ...testEnv.spec,
+        displayName: undefined,
+      },
+      healthStatus: 'Healthy',
+    };
+    render(<EnvironmentCard environment={env} />);
+    expect(screen.getByText('Manual')).toBeVisible();
+    expect(screen.getByText('Healthy')).toBeVisible();
   });
 });
