+++ src/components/Environment/create/__tests__/CreateEnvironment.spec.tsx
@@ -99,7 +99,7 @@ describe('CreateEnvironment', () => {
     const submitButton = screen.getByRole('button', { name: 'Create environment' });
     fireEvent.click(submitButton);
     await waitFor(() =>
-      expect(navigateMock).toHaveBeenCalledWith('/stonesoup/workspaces/test-ws/workspace-settings'),
+      expect(navigateMock).toHaveBeenCalledWith('/stonesoup/environments/workspaces/test-ws'),
     );
   });
 
@@ -107,6 +107,6 @@ describe('CreateEnvironment', () => {
     render(<CreateEnvironment />);
     const cancelButton = screen.getByText('Cancel');
     fireEvent.click(cancelButton);
-    expect(navigateMock).toHaveBeenCalledWith('/stonesoup/workspaces/test-ws/workspace-settings');
+    expect(navigateMock).toHaveBeenCalledWith('/stonesoup/environments/workspaces/test-ws');
   });
 });
