+++ src/pages/__tests__/EnvironmentsListPage.spec.tsx
@@ -1,11 +1,10 @@
 import * as React from 'react';
 import '@testing-library/jest-dom';
-import { useFeatureFlag } from '@openshift/dynamic-plugin-sdk';
 import { useK8sWatchResource } from '@openshift/dynamic-plugin-sdk-utils';
 import { configure, screen } from '@testing-library/react';
 import { useAccessReviewForModels } from '../../utils/rbac';
 import { namespaceRenderer } from '../../utils/test-utils';
-import WorkspaceSettingsPage from '../WorkspaceSettingsPage';
+import EnvironmentsListPage from '../EnvironmentsListPage';
 
 jest.mock('react-router-dom', () => {
   const actual = jest.requireActual('react-router-dom');
@@ -20,27 +19,27 @@ jest.mock('@openshift/dynamic-plugin-sdk-utils', () => ({
   getActiveWorkspace: jest.fn(() => 'test-ws'),
 }));
 
-jest.mock('@openshift/dynamic-plugin-sdk', () => ({
-  useFeatureFlag: jest.fn(),
+jest.mock('@redhat-cloud-services/frontend-components/useChrome', () => ({
+  useChrome: () => ({
+    helpTopics: { setActiveTopic: jest.fn(), enableTopics: jest.fn(), disableTopics: jest.fn() },
+  }),
 }));
 
 jest.mock('../../utils/rbac', () => ({
   useAccessReviewForModels: jest.fn(),
 }));
 
-jest.mock('../../components/WorkspaceSettings/WorkspaceSettings', () => () => {
-  return <div data-test="workspace-settings-page" />;
+jest.mock('../../components/Environment/EnvironmentListView', () => () => {
+  return <div data-test="environments-list-view" />;
 });
 
 const accessReviewMock = useAccessReviewForModels as jest.Mock;
 const watchResourceMock = useK8sWatchResource as jest.Mock;
-const useFeatureFlagMock = useFeatureFlag as jest.Mock;
 
 configure({ testIdAttribute: 'data-test' });
 
-describe('WorkspaceSettingsPage', () => {
+describe('EnvironmentsListPage', () => {
   beforeEach(() => {
-    useFeatureFlagMock.mockReturnValue([false]);
     accessReviewMock.mockReturnValue([true, true]);
   });
 
@@ -51,16 +50,16 @@ describe('WorkspaceSettingsPage', () => {
   it('should show the spinner when access check is not loaded', () => {
     accessReviewMock.mockReturnValue([true, false]);
     watchResourceMock.mockReturnValue([[], true, null]);
-    namespaceRenderer(<WorkspaceSettingsPage />, 'test-ns', { workspacesLoaded: true });
+    namespaceRenderer(<EnvironmentsListPage />, 'test-ns', { workspacesLoaded: true });
+    screen.debug();
 
     screen.getByTestId('spinner');
   });
 
-  it('should render workspace settings page', () => {
+  it('should render environments list page', () => {
     accessReviewMock.mockReturnValue([true, true]);
     watchResourceMock.mockReturnValue([[], true, null]);
-    namespaceRenderer(<WorkspaceSettingsPage />, 'test-ns', { workspacesLoaded: true });
-
-    screen.getByTestId('workspace-settings-page');
+    namespaceRenderer(<EnvironmentsListPage />, 'test-ns', { workspacesLoaded: true });
+    screen.getByTestId('environments-list-view');
   });
 });
